Restore "toot" flavor in our custom locale
I think it was a mistake to remove this quirky little word from Mastodon
mainline. It gives a little bit of character to the software over the
lifeless replacement of "post" and I'm not convinced that the argument
of "making Mastodon a better citizen of the Fediverse" actually holds
any merit. A lot of other instances already call their posts "growls"
or "awoos" or some other noise, but there's not exactly a sound of
plurality that we could borrow for local flavor. So we're sticking with
"toot" for the time being.

I'm not interested in maintaining *every* locale, but conveniently we
already have one set up to customize the name of the local timeline to
"Plural Café" so we can use that one.