Transition away from GitHub repository
* Change link footer and onboarding model to point to our personal Git
  repository and direct folks to email us for collaboration
* Remove now-unnecessary .github folder and add it to .gitignore [I am
  not actually sure if this will prevent us from getting those files
  again in the future.]
* Restore English locale to upstream version and move our changes into
  the en-cafe custom locale.
* Remove most of the README for clarity.

After GitHub's contract with ICE came to light, we moved our personal
projects away from the site. With Copilot now effectively laundering
code under free licenses through machine learning by claiming that
training a model is "fair use" even when it directly quotes the input,
it's clear now that making the effort to move our Mastodon fork off of
GitHub is worth the effort. But even if the model *never* recited
anyone's source, we consider it to be clearly unethical to use open
source code in this way without a permissive license or written
consent from the authors of any non-trivial contributions. Whether or
not the law agrees, this is clearly a derivative work in the colloquial
sense of the term.

We needed to build our own CI/CD solution in the next week and a half
anyway because Docker Hub is removing automatic builds from the free
tier.

We're under no illusions that moving to our personal Git host will
actually stop anyone from doing whatever they want with what we post
publicly on the Internet. But we're both sending a message and
learning how to build our own infrastructure, which will hopefully be
helpful to us in the future.

You can read more about GitHub's ICE deals here:
https://www.zdnet.com/article/microsoft-github-ceo-why-we-defend-ice-deal-in-the-face-of-employee-anger/

You can read about GitHub Copilot quoting source code, and the steps
they've taken to reduce its frequency, here:
https://docs.github.com/en/github/copilot/research-recitation